Image Classifier to Identify Dog Breeds
Use a Pre-trained Image Classifier to Identify Dog Breeds
Classifies pet images using a pre-trained CNN model, compares these classifications to the true identity of the pets in the images, and summarizes how well the CNN performed on the image classification task. Therefore, the program must first extract the pet image label from the filename before classifying the images using the pre-trained CNN model. With this program, it will be comparing the performance of 3 different CNN model architectures to determine which provides the 'best' classification.